requirements.txt
# Core framework
fastapi==0.115.5
uvicorn==0.32.1
starlette==0.40.0

# Data validation & settings
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.6.1

# Common dependencies
python-dotenv==1.0.1
requests==2.32.3
PyYAML==6.0.2
anyio==4.6.2.post1
typing-extensions>=4.12.2

# (Optional) Useful for debugging & CORS
aiofiles==24.1.0
fastapi[all]==0.115.5
python-jose==3.3.0
passlib[bcrypt]==1.7.4

httpx
xmltodict
python-dotenv

gunicorn==23.0.0
zeep==4.3.2
